% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Block bootstrap}
\usage{
block_bootstrap(df, m, block, .valid.pred)
}
\arguments{
\item{.valid.pred}{}
}
\description{
From https://cran.r-project.org/web/packages/broom/vignettes/bootstrapping.html
}

