% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap_}
\alias{block_bootstrap_}
\title{Block bootstrap}
\usage{
block_bootstrap_(df, m, block, .valid.pred)
}
\arguments{
\item{df}{data to bootstrap}

\item{m}{number of samples}

\item{block}{string vector of IDs identifying blocks}

\item{.valid.pred}{predicate function to accept samples}
}
\description{
From https://cran.r-project.org/web/packages/broom/vignettes/bootstrapping.html
}
