% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_lm.R
\name{predict.lm_strat}
\alias{predict.lm_strat}
\title{Predict outcomes for stratified linear regressions}
\usage{
\method{predict}{lm_strat}(fm, newdata, ...)
}
\arguments{
\item{...}{}
}
\description{
Predict outcomes for stratified linear regressions
}
