% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_lm.R
\name{r_squared}
\alias{r_squared}
\title{Calculate the R Squared for stratified regressions}
\usage{
r_squared(fm, adjusted = TRUE)
}
\arguments{
\item{fm}{stratified regression results}

\item{adjusted}{}
}
\description{
Calculate the R Squared for stratified regressions
}
