% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_lm.R
\name{run_strat_reg.default}
\alias{run_strat_reg.default}
\title{Run Stratified Regression}
\usage{
\method{run_strat_reg}{default}(
  .data,
  .formula,
  .cluster,
  .strat.by = NULL,
  .covariates = NULL,
  ...
)
}
\arguments{
\item{.covariates}{}
}
\description{
Run Stratified Regression
}
