% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_lm.R
\name{tidy.lm_strat}
\alias{tidy.lm_strat}
\title{Tidy stratified regression results}
\usage{
\method{tidy}{lm_strat}(fm, .include_covar = FALSE, ...)
}
\arguments{
\item{.include_covar}{}

\item{...}{}
}
\description{
Tidy stratified regression results
}
