% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_robust.R
\name{vcov_clx.default}
\alias{vcov_clx.default}
\title{Calculate the Cluster Robust Variance Covariance Matrix}
\usage{
\method{vcov_clx}{default}(fm, cluster, ...)
}
\arguments{
\item{cluster}{}
}
\description{
Taken from ivpack::clx()
}
