% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Block bootstrap}
\usage{
block_bootstrap(df, m, ..., .valid.pred)
}
\arguments{
\item{df}{data to bootstrap}

\item{m}{number of samples}

\item{...}{IDs identifying blocks}

\item{.valid.pred}{predicate function to accept samples}
}
\description{
Block bootstrap
}
